moved ExtendingClassTest
