+++ packages/Reflection/tests/Parser/ExtendingSources/ExtendingClass.php
@@ -1,0 +1,9 @@
+<?php declare(strict_types=1);
+
+namespace ApiGen\Reflection\Tests\Parser\ExtendingSources;
+
+use ApiGen\Reflection\Tests\Parser\NotLoadedSources;
+
+class ExtendingClass extends NotLoadedSources\SomeClass
+{
+}
