+++ packages/Reflection/tests/Parser/NotLoadedSources/SomeClass.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace ApiGen\Reflection\Tests\Parser\NotLoadedSources;
+
+class SomeClass
+{
+}
