--- tests/Generator/ExtendingSources/ExtendingClass.php
@@ -1,10 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace ApiGen\Tests\Generator\ExtendingSources;
-
-use ApiGen\Tests\Generator\NotLoadedSources;
-
-class ExtendingClass extends NotLoadedSources\SomeClass
-{
-
-}
