--- tests/Generator/NotLoadedSources/SomeClass.php
@@ -1,8 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace ApiGen\Tests\Generator\NotLoadedSources;
-
-class SomeClass
-{
-
-}
